COMP-267: Scenario 3 Selecting ‘EAC Amendment’ as a Requirement Source
