+++ compliance-web/src/components/App/Complaints/Profile/ComplaintGeneralInformation.tsx
@@ -1,10 +1,8 @@
-import {
-  ComplaintSourceEnum,
-  RequirementSourceEnum,
-} from "@/components/App/Complaints/ComplaintFormUtils";
+import { ComplaintSourceEnum } from "@/components/App/Complaints/ComplaintFormUtils";
 import FileProfileProperty from "@/components/App/FileProfileProperty";
 import { Complaint } from "@/models/Complaint";
 import { useMenuStore } from "@/store/menuStore";
+import { RequirementSourceEnum } from "@/utils/constants";
 import dateUtils from "@/utils/dateUtils";
 import { EditRounded } from "@mui/icons-material";
 import { Box, Button, Typography } from "@mui/material";
@@ -22,7 +20,11 @@ const ComplaintGeneralInformation: React.FC<
   const { appHeaderHeight } = useMenuStore();
 
   const generalProperties = [
-    { name: "Case File", value: complaintData.case_file.case_file_number, link: true},
+    {
+      name: "Case File",
+      value: complaintData.case_file.case_file_number,
+      link: true,
+    },
     { name: "Project Name", value: complaintData.case_file?.project?.name },
     {
       name: "Concern Description",
@@ -158,8 +160,8 @@ const ComplaintGeneralInformation: React.FC<
             key={property.name}
             propertyName={property.name}
             propertyValue={property.value}
-            linksList={property.link ? [property.value] : null}
-            linkRoute={property.link ? "/ce-database/case-files" : null}
+            linksList={property.link ? [property.value] : undefined}
+            linkRoute={property.link ? "/ce-database/case-files" : undefined}
           />
         ))}
       </Box>
